Rename JS variation block support to block style variation
