+++ packages/block-editor/src/hooks/index.js
@@ -24,7 +24,7 @@ import fontSize from './font-size';
 import textAlign from './text-align';
 import border from './border';
 import position from './position';
-import variation from './variation';
+import blockStyleVariation from './block-style-variation';
 import layout from './layout';
 import childLayout from './layout-child';
 import contentLockUI from './content-lock-ui';
@@ -62,7 +62,7 @@ createBlockListBlockFilter( [
 	fontSize,
 	border,
 	position,
-	variation,
+	blockStyleVariation,
 	childLayout,
 ] );
 createBlockSaveFilter( [
